backend/requirements.txt
# FastAPI核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.13.1

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0

# 文件上传和表单处理
python-multipart==0.0.6

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 配置管理
python-dotenv==1.0.0

# HTTP客户端
httpx==0.25.2

# 图片处理
pillow==10.1.0

# 时间处理
python-dateutil==2.8.2

# 日志
loguru==0.7.2

# PDF生成
reportlab==4.0.7

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
